+++ lib/todoro/engine.rb
@@ -7,5 +7,9 @@ module Todoro
         extend Todoro::Taskable::ClassMethods
       end
     end
+
+    initializer "todoro.assets.precompile" do |app|
+      app.config.assets.precompile += %w[todoro/application.css todoro/application.js]
+    end
   end
 end