+++ app/models/todoro/task_step.rb
@@ -1,0 +1,7 @@
+module Todoro
+  class TaskStep < ApplicationRecord
+    belongs_to :task, class_name: "Todoro::Task"
+
+    validates :title, presence: true
+  end
+end